requirements.txt
selenium==4.5.0
pymongo==4.4.0
requests==2.28.1
loguru==0.6.0
mysql-connector-python==8.0.33

